Adding a Bloodshed project for creating a DLL version of the framework.  It is not yet tested, but it does build.
Silencing various warnings.

Killing off pattr support on Windows, though it will still work on the Mac.


git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@2877 464221e2-0a0f-0410-983b-f8d57d210c17

+++ SourceCode/Framework/DataspaceLib/TimeDataspace.cpp
@@ -37,7 +37,7 @@ SampleUnit::SampleUnit()
 	: DataspaceUnit::DataspaceUnit("sample")
 {
 	// TODO: Do something a bit more legit with the sample rate -- need to actually know the real sample rate!
-	sample_rate = DEFAULT_SAMPLE_RATE;
+	sample_rate = (long)DEFAULT_SAMPLE_RATE;
 	msr = double(sample_rate) / 1000.0;
 	rmsr = 1.0 / msr;
 }
